requirements.txt
requests==2.28.2
colorama==0.4.6
python-dotenv==0.19.2
psycopg2-binary==2.9.5
